software/requirements.txt
numpy
opencv-python
rerun-sdk
sounddevice
dashscope
scipy
flask
pyzmq
ruff; extra == 'dev'
